pyproject.toml
[project]
name = "newsletter-mcp"
version = "0.1.0"
description = "MCP server for ingesting Smart Spreads newsletter PDFs into a database."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.18.0",
  "pydantic>=2.9.0",
  "pypdf>=6.10.1",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
]

[project.scripts]
newsletter-mcp = "newsletter_mcp.server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
